showing warning for embedded players
